fix: use pnpm workspace to ensure the typings are installed
Fixes the vite-plugin-vue issue while still not having to break the
language-tools test.
But this is only a temporary solution.
Because the failure in vite-plugin-vue tests indicates that the root
tsconfig and dependencies still somehow affects the workspace typings.
In the long run we need to refactor the folder structure to avoid any
possible interferences.